backend/requirements.txt
fastapi
uvicorn
gunicorn
pydantic-ai
google-generativeai
python-dotenv
PyMuPDF
pytesseract
pdf2image
Pillow